Show selected value in browse page buttons
