+++ src/utils/utils.js
@@ -70,6 +70,17 @@ const curateFilters = [
   { value: 'described', label: 'Focus on description issues' }
 ]
 
+const providers = [
+  { value: 'cocoapods', label: 'CocoaPods' },
+  { value: 'cratesio', label: 'Crates.io' },
+  { value: 'github', label: 'GitHub' },
+  { value: 'mavencentral', label: 'MavenCentral' },
+  { value: 'npmjs', label: 'NpmJS' },
+  { value: 'nuget', label: 'NuGet' },
+  { value: 'pypi', label: 'PyPi' },
+  { value: 'rubygems', label: 'RubyGems' }
+]
+
 const noRowsHeight = 200
 
 export {
@@ -79,6 +90,7 @@ export {
   difference,
   licenses,
   noRowsHeight,
+  providers,
   releaseDates,
   setIfValue,
   sorts,
